backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
requests==2.31.0
python-dotenv==1.0.1
google-generativeai>=0.7.0
scikit-learn>=1.3.0
xgboost>=2.0.0
pandas>=2.1.0
numpy>=1.26.0
PyGithub>=2.1.1
